Dtr Online Bug Fix

+++ src/main/java/com/sericulture/model/DTROnlineReportResponse.java
@@ -18,10 +18,14 @@ public class DTROnlineReportResponse extends ResponseBody {
     private List<DTROnlineReportUnitDetail> dtrOnlineReportUnitDetailList = new ArrayList<>();
     private int totalLots;
     private int paymentSuccessLots;
-    private double totalFarmerAmount;
-    private double totalReelerAmount;
-    private double totalReelerMarketFee;
-    private double totalFarmerMarketFee;
+    //    private double totalFarmerAmount;
+//    private double totalReelerAmount;
+//    private double totalReelerMarketFee;
+//    private double totalFarmerMarketFee;
+    private float totalFarmerAmount;
+    private float totalReelerAmount;
+    private float totalReelerMarketFee;
+    private float totalFarmerMarketFee;
 
     private int totalBidAmount;
     private float totalWeight;
